% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbClean.R
\name{nbClean}
\alias{nbClean}
\title{Clean dataset}
\usage{
nbClean(data = NULL)
}
\arguments{
\item{data}{Datasettet som er ferdig konvertert til R format}
}
\value{
Renset datasettet til videre behandling
}
\description{
Denne funksjonen skal filtrere datasettet som er spesifisert i datautvalget
}

