% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbFigur.R
\name{nbFigur}
\alias{nbFigur}
\title{Valgt av sykehus og rapport}
\usage{
nbFigur(data = NULL, sykehus = NULL, rapportValg = NULL, yAksen = NULL,
  save = NULL, format = "pdf", path = NULL)
}
\arguments{
\item{data}{Datasettet}

\item{sykehus}{Utvalgte sykehus}

\item{rapportValg}{Rapport valg}

\item{yAksen}{Y aksen for figuren}

\item{save}{Filnavn for å lage figuren}

\item{format}{Figur format i.e "pdf", "jpg" eller "png"}

\item{path}{Hvor skal figuren lagres evt. hjemme område}
}
\description{
Her skal deg gjøres valg av sykehus for sammenligne og type rapport
}

