% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbFile.R
\name{nbFile}
\alias{nbFile}
\title{Konvertering datasettet til R format}
\usage{
nbFile(filnavn = NULL)
}
\arguments{
\item{filnavn}{Filnavn hentet fra register database}
}
\value{
Et dataset i R data.frame format
}
\description{
Denne funksjon skal konvertere datafil fra SPSS eller CSV format
til en format som R kan behandle
}

