% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbFilter.R
\name{nbFilter}
\alias{nbFilter}
\title{Filtrer datasett til videre analyser}
\usage{
nbFilter(minAlder = NULL, maxAlder = NULL, datoFra = NULL,
  datoTil = NULL, kjonn = NULL, dbType = NULL, dataValg = NULL)
}
\arguments{
\item{minAlder}{Minimum alder}

\item{maxAlder}{Maksimum alder}

\item{datoFra}{Tidligste dato i utvalget}

\item{datoTil}{Seneste dato i utvalget}

\item{kjonn}{Kjønn}

\item{dbType}{Diabetes type}

\item{dataValg}{Valg av data f.eks. Førstegangsregister, Årskontroll etc}
}
\value{
List datasett
}
\description{
Produserer et datasett basert på utvalgte parametrer før videre analyser
eller visualisering.
}
\details{
Denne funksjonen produserer disse datasetter:
\itemize{
 \item \code{fileSelect} {Utvalgte parametre for filtrering}
 \item \code{*vec} {Vektor for utvalgte parametrene}
}
}
\note{
\code{dbFile} og \code{dbClean} funksjoner er kjørt før datasettet
blir brukt på denne funksjonen
}

