% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbSelectVar.R
\name{nbSelectVar}
\alias{nbSelectVar}
\title{Velger variabel til figur}
\usage{
nbSelectVar(data = NULL, valgtVar = NULL)
}
\arguments{
\item{data}{Datasettet}

\item{valgtVar}{Utvalgte variabel til figuren}
}
\description{
Her skal utvalgte variabel videre behandlet for figuren er lages
}

