% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbText.R
\name{nbText}
\alias{nbText}
\title{Lager tekst til figur}
\usage{
nbText(data = NULL, fdata = NULL, minAlder = NULL, maxAlder = NULL,
  datoFra = NULL, datoTil = NULL, kjonn = NULL, dbType = NULL,
  dataValg = NULL)
}
\arguments{
\item{data}{Datasettet som er ferdig konvertert til R format}

\item{fdata}{Filtrerings parametre}

\item{minAlder}{Minimum alder}

\item{maxAlder}{Maksimum alder}

\item{datoFra}{Tidligste dato i utvalget}

\item{datoTil}{Seneste dato i utvalget}

\item{kjonn}{Kjønn}

\item{dbType}{Diabetes type}

\item{dataValg}{Valg av data f.eks. Førstegangsregister, Årskontroll etc}
}
\value{
Rensert datasett og tekst til figur
}
\description{
Denne funksjonen skal lage tekst til figur basert fra datasettet som allerede er filtret
}

