% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumCI.R
\name{sumCI}
\alias{sumCI}
\title{Bregne Konfidens Interval}
\usage{
sumCI(data, maalvar, gpvars, conf.Int = 0.95)
}
\arguments{
\item{data}{Datasettet}

\item{maalvar}{Utvalgte variabel}

\item{gpvars}{Gruppe}

\item{conf.Int}{Konfidence nivå f.eks .95}
}
\description{
Funksjonen brukes til å bregne Konfidens Interval (CI)
}

